+++ dotCMS/src/main/java/com/dotcms/rest/api/v1/content/_import/ContentImportForm.java
@@ -1,4 +1,4 @@
-package com.dotcms.rest.api.v1.contentImport;
+package com.dotcms.rest.api.v1.content._import;
 
 import com.dotcms.repackage.javax.validation.constraints.NotNull;
 import com.dotcms.rest.api.Validated;
